rbp_backend/requirements.txt
flask
flask-cors
python-dotenv
requests
PyJWT
supabase
psycopg2-binary
stripe
google-generativeai
gunicorn
